% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Australia.R
\docType{data}
\name{Australia}
\alias{Australia}
\title{A map of Australia}
\format{A SpatialPolygonsDataFrame}
\usage{
Australia
}
\description{
This map layer outlines the coast of Australia.
}
\keyword{datasets}
