% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap.R
\docType{data}
\name{basemap}
\alias{basemap}
\title{A map of Sandy Strait, Australia}
\format{A SpatialPolygonsDataFrame}
\usage{
basemap
}
\description{
This map layer outlines the coast around Sandy Strait, Australia.
}
\keyword{datasets}
