% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curtis.R
\docType{data}
\name{curtis}
\alias{curtis}
\title{Flatback turtle utilisation distributions}
\format{A list of 29 RasterLayer objects}
\source{
Shimada T et al, (in prep)
}
\usage{
curtis
}
\description{
Inter-nesting utilisation distributions of 29 flatback turtles (\emph{Natator depressus}) that nested in Curtis Island, Australia.
As a demonstration, the resolution of the UDs is reduced to 1km grid instead of 50m as used in Shimada et al (in prep).
}
\keyword{datasets}
