% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupfilter.space.r
\name{dupfilter.space}
\alias{dupfilter.space}
\title{Filter spatial duplicates}
\usage{
dupfilter.space(sdata, step.time = 0, step.dist = 0,
  conditional = FALSE)
}
\arguments{
\item{sdata}{A data frame containing columns with the following headers: "id", "DateTime", "lat", "lon", "qi". 
This filter is independently applied to a subset of data grouped by the unique "id". "DateTime" is date & time in class \code{\link[base]{POSIXct}}. 
"lat" and "lon" are the recorded latitude and longitude in decimal degrees. 
"qi" is the numerical quality index associated with each fix where the greater number represents better quality 
(e.g. number of GPS satellites used for estimation).}

\item{step.time}{A numeric value specifying temporal interval between two consecutive locations. Default is 0 hours. 
Locations are considered temporal duplicates if the temporal interval is less than or equal to the user specified value.}

\item{step.dist}{A numeric value specifying spatial interval between two consecutive locations. Default is 0 kilometres. 
Locations are considered spatial duplicates if the spatial interval is less than or equal to the user specified value.}

\item{conditional}{If TRUE, spatial duplicates are removed only if the temporal interval between the locations is 
less than the time specified in "step.time". Default is FALSE.}
}
\value{
Input data frame is returned with spatial duplicates removed. 
The following columns are added: "pTime", "pDist". 
"pTime" is hours from a previous fix. "pDist" is straight distance in kilometres from a previous fix. 
#' @author Takahiro Shimada
}
\description{
A partial component of dupfilter although works as a stand-alone function. This function removes spatial duplicates.
}
\details{
This function selects a fix from multiple fixes which were obtained at the same geographical coordinate. 
A minimum of two locations per id is required to run this function.
}
\references{
Shimada T, Limpus C, Jones R, Hazel J, Groom R, Hamann M (2016) 
Sea turtles return home after intentional displacement from coastal foraging areas. 
Marine Biology 163:1-14 doi:10.1007/s00227-015-2771-0
}
\seealso{
\code{\link{dupfilter}}, \code{\link{dupfilter.exact}}, \code{\link{dupfilter.time}}, \code{\link{dupfilter.qi}}
}
