% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupfilter_exact.R
\name{dupfilter_exact}
\alias{dupfilter_exact}
\title{Filter temporally and spatially exact duplicates}
\usage{
dupfilter_exact(sdata)
}
\arguments{
\item{sdata}{A data frame containing columns with the following headers: "id", "DateTime", "lat", "lon", "qi". 
The function filters the input data by the unique "id". 
"DateTime" is date & time in class \code{\link[base]{POSIXct}}. 
"lat" and "lon" are the latitude and longitude of each location in decimal degrees. 
"qi" is the numerical quality index associated with each location fix where the greater number indicates the higher accuracy 
(e.g. the number of GPS satellites involved in estimation).}
}
\value{
The input data frame is returned with temporally and spatially exact duplicates removed. 
The following columns are added: "pTime", "sTime", "pDist", "sDist". 
"pTime" and "sTime" are hours from a previous and to a subsequent fix respectively. 
"pDist" and "sDist" are straight distances in kilometres from a previous and to a subsequent fix respectively.
}
\description{
Function to filter temporally and spatially exact locations in tracking data.
}
\details{
This is a partial component of \code{\link{dupfilter}}, although works as a stand-alone function.
It looks for temporally and spatially exact locations and retains only a single fix (latitude/longitude pair) per time and location.
}
\references{
Shimada T, Limpus C, Jones R, Hazel J, Groom R, Hamann M (2016) 
Sea turtles return home after intentional displacement from coastal foraging areas. 
\emph{Marine Biology} 163:1-14 doi:\href{http://doi.org/10.1007/s00227-015-2771-0}{10.1007/s00227-015-2771-0}
}
\seealso{
\code{\link{dupfilter}}, \code{\link{dupfilter_qi}}, \code{\link{dupfilter_time}}, \code{\link{dupfilter_space}}
}
\author{
Takahiro Shimada
}
