% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupfilter_space.R
\name{dupfilter_space}
\alias{dupfilter_space}
\title{Filter spatial duplicates}
\usage{
dupfilter_space(sdata, step.time = 0, step.dist = 0, conditional = FALSE)
}
\arguments{
\item{sdata}{A data frame containing columns with the following headers: "id", "DateTime", "lat", "lon", "qi". 
The function filters the input data by the unique "id". 
"DateTime" is date & time in class \code{\link[base]{POSIXct}}. 
"lat" and "lon" are the latitude and longitude of each location in decimal degrees. 
"qi" is the numerical quality index associated with each location fix where a greater number indicates a higher accuracy 
(e.g. the number of GPS satellites involved in estimation).}

\item{step.time}{Consecutive locations less than or equal to \emph{step.time} apart are considered temporal duplicates.
Default is 0 hours.}

\item{step.dist}{Consecutive locations less than or equal to \emph{step.dist} apart are considered spatial duplicates.
Default is 0 kilometres.}

\item{conditional}{If TRUE, spatial duplicates are filtered only if they are less than or equal to \emph{step.time} apart. 
Default is FALSE.}
}
\value{
The input data frame is returned with spatial duplicates removed. 
The following columns are added: "pTime", "sTime", "pDist", "sDist". 
"pTime" and "sTime" are hours from a previous and to a subsequent fix respectively. 
"pDist" and "sDist" are straight distances in kilometres from a previous and to a subsequent fix respectively.
}
\description{
Function to filter spatial duplicates in tracking data.
}
\details{
This function is a partial component of \code{\link{dupfilter}}, although works as a stand-alone function.
First it identifies spatial duplicates by searching for consecutive fixes that were located within \emph{step.dist}.
For each group of spatial duplicates, the function then retains a single fix that is nearest from a previous and to a subsequent location.
}
\note{
A minimum of two locations per id is required.
}
\references{
Shimada T, Limpus C, Jones R, Hazel J, Groom R, Hamann M (2016) 
Sea turtles return home after intentional displacement from coastal foraging areas. 
\emph{Marine Biology} 163:1-14 doi:\href{http://doi.org/10.1007/s00227-015-2771-0}{10.1007/s00227-015-2771-0}
}
\seealso{
\code{\link{dupfilter}}, \code{\link{dupfilter_exact}}, \code{\link{dupfilter_time}}, \code{\link{dupfilter_qi}}
}
\author{
Takahiro Shimada
}
