% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kml_track.R
\name{kml_track}
\alias{kml_track}
\title{Generate KML from locations}
\usage{
kml_track(sdata, crs = 4326, output = "open", type = "point", ...)
}
\arguments{
\item{sdata}{A data frame containing location data of \bold{one} individual, 
with the following column headers: "id", "DateTime", "lat", "lon". 
"id" is an identifier of the individual. 
"DateTime" is the GMT date & time of each location in class \code{\link[base:DateTimeClasses]{POSIXct}} 
or \code{\link[base]{character}} with the following format "2012-06-03 01:33:46".
"lat" and "lon" are the latitude and longitude of each location in decimal degrees.}

\item{crs}{A number specifying the European Petroleum Survey Group (EPSG) code for the input location data.}

\item{output}{A string specifying whether to 'open' or 'save' the output. 
The output will be saved in the current working directory.}

\item{type}{Type of the output. 'point' or 'line.}

\item{...}{Optional arguments passed to \code{\link[plotKML]{plotKML}} and \code{\link[plotKML]{kml}}.}
}
\value{
A kml file
}
\description{
Function to generate a kml file from tracking data. 
This is a wrapper of \code{\link[plotKML]{plotKML}} and \code{\link[plotKML]{kml}}, 
specifically designed to generate a kml file from tracking data.
}
\examples{
#### Fastloc GPS data obtained from a green turtle
data(turtle)

\dontrun{
#### See the data on Google earth
## points with time stamps
kml_track(turtle, output = 'open', type = 'point', 
points_names = turtle$DateTime, colour_scale = 'yellow')

## lines
kml_track(turtle, output = 'open', type = 'line', colour = 'red')
 

#### Save the location points to the current working directory    
shape <- "http://maps.google.com/mapfiles/kml/pal2/icon26.png"
kml_track(turtle, output = 'save', type = 'point', shape = shape, colour = 'yellow')
}
}
\seealso{
\code{\link{map_track}}
}
\author{
Takahiro Shimada
}
