% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_vol.R
\name{percent_vol}
\alias{percent_vol}
\title{UD percent volume}
\usage{
percent_vol(x, percent = 100)
}
\arguments{
\item{x}{A vector containing the probability density.}

\item{percent}{An integer specifying the percent volume of a UD to be considered.}
}
\value{
A vector containing the specified percent volume.
}
\description{
Function to calculate a percent volume on a utilisation distribution (UD)
}
\details{
This function calculates a percent volume on a UD. The probability beyond the specified range will be assigned with a zero value.
}
\author{
Takahiro Shimada
}
