% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandyStrait.R
\docType{data}
\name{SandyStrait}
\alias{SandyStrait}
\title{A map of Sandy Strait, Australia}
\format{A SpatialPolygonsDataFrame}
\usage{
SandyStrait
}
\description{
This map layer outlines the coast around Sandy Strait, Australia.
}
\keyword{datasets}
