% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle.R
\docType{data}
\name{turtle}
\alias{turtle}
\title{Green turtle tracking data}
\format{A data frame with 429 rows and 5 variables:
\describe{
  \item{id}{identifier for each animal.}
  \item{DateTime}{date & time in class \code{\link[base]{POSIXct}}.}
  \item{lat}{latitude in decimal degrees.}
  \item{lon}{longitude in decimal degrees.}
  \item{qi}{numerical quality index associated with each location fix where the greater number indicates the higher accuracy
   (e.g. number of GPS satellites used for estimation).}
}}
\source{
Shimada T, Jones R, Limpus C, Groom R, Hamann M (2016) 
Long-term and seasonal patterns of sea turtle home ranges in warm coastal foraging habitats: Implications for conservation. 
\emph{Marine Ecology Progress Series} 562:163-179. doi:\href{http://doi.org/10.3354/meps11972}{10.3354/meps11972}
}
\usage{
turtle
}
\description{
A dataset containing Fastloc GPS locations of a green turtle tracked in Sandy Strait, Australia.
}
\keyword{datasets}
