% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle2.R
\docType{data}
\name{turtle2}
\alias{turtle2}
\title{Green turtle tracking data 2}
\format{
A data frame with 276 rows and 5 variables:
\describe{
  \item{id}{identifier for each animal.}
  \item{DateTime}{GMT date & time of each location in class \code{\link[base:DateTimeClasses]{POSIXct}}.}
  \item{lat}{latitude in decimal degrees.}
  \item{lon}{longitude in decimal degrees.}
  \item{qi}{quality index associated with each location fix. 
The input values can be either the number of GPS satellites or Argos Location Classes. 
Argos Location Classes will be converted to numerical values, where "A", "B", "Z" will be replaced with "-1", "-2", "-3" respectively.
The greater number indicates a higher accuracy.}
}
}
\source{
Shimada T, Jones R, Limpus C, Groom R, Hamann M (2016) 
Long-term and seasonal patterns of sea turtle home ranges in warm coastal foraging habitats: Implications for conservation. 
\emph{Marine Ecology Progress Series} 562:163-179. \doi{10.3354/meps11972}
}
\usage{
turtle2
}
\description{
A dataset containing Fastloc GPS locations of a green turtle tracked in Moreton Bay, Australia.
}
\keyword{datasets}
