% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_area.R
\name{boot_area}
\alias{boot_area}
\title{Cumulative analysis of collective areas by bootstrapping}
\usage{
boot_area(
  data,
  cell.size = NA,
  R = 1000,
  percent = 50,
  quantiles = c(0.25, 0.5, 0.75)
)
}
\arguments{
\item{data}{A matrix or list of RasterLayer objects. 
Each row of the matrix or each RasterLayer object contains a utilisation distribution 
(or other statistics that sums to 1 - e.g. proportion of time spent).
\bold{The grid size and geographical extent must be consistent across each row of the matrix or each RasterLayer object.}
The function assumes that each column of the matrix is associated with a unique geographical location or 
that each RasterLayer has exactly the same geographical extent and resolution.}

\item{cell.size}{A numeric value specifying the grid cell size of the input data in metres.}

\item{R}{An integer specifying the number of iterations.}

\item{percent}{An integer specifying the percent volume of each UD to be considered in the analysis.}

\item{quantiles}{A vector or a number to specify the quantiles to be calculated in the summary of the results.}
}
\value{
A list containing two data frames - raw results and summary (mean, sd, sem and quantiles at each sample size).
}
\description{
Function to calculate collective areas (merged x\% Utilisation Distributions or UDs) of n individuals by bootstrapping.
}
\details{
This function calculates collective areas (e.g. 50\% UDs) of 1 to n individuals by bootstrapping.
}
\examples{
\dontrun{

#1 Utilisation distributions of flatback turtles.
data(curtis)

#2 Calculate collective areas from 1000 random permutation
area <- boot_area(curtis, R = 1000, percent = 50)

#3 Find the minimum sample size required to estimate the general distribution.
a <- asymptote(area)

#4 Plot the mean collective area and rational function fit relative to the sample sizes.
ggplot(data = area$summary)+
  geom_point(aes(x = N, y = mu/1e+6), alpha = 0.5) + 
  geom_path(data = a$results, aes(x = x, y = ys/1e+6)) +        
  labs(x = "N", y = expression(Area~(km^2)))
}
}
\references{
Shimada T, Thums M, Hamann M, Limpus CJ, Hays GC, FitzSimmons N, Wildermann NE, Duarte CD, Meekan MG (in press) 
Optimising sample sizes for animal distribution analysis using tracking data. 
\emph{Methods in Ecology and Evolution} n/a(n/a) doi:\href{http://doi.org/10.1111/2041-210X.13506}{10.1111/2041-210X.13506}
}
\author{
Takahiro Shimada
}
