% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidedata.R
\docType{data}
\name{tidedata}
\alias{tidedata}
\title{Tidal data for Sandy Strait, Australia}
\format{A data frame with 26351 rows and 3 variables:
\describe{
  \item{tideDT}{date & time in class POSIXct for each observed tidal height}
  \item{reading}{observed tidal height in meters}
  \item{standard.port}{identifier of the tidal station}
}}
\source{
The State of Queensland (Department of Transport and Main Roads), Tidal data.
}
\usage{
tidedata
}
\description{
A dataset containing tidal observations recorded at Bundaberg, Australia
}
\keyword{datasets}
